+++ src/timePicker/index.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { color, inputBorderSize, font } from '_utils/branding'
+import { color, inputBorderSize, font, space } from '_utils/branding'
 
 import TimePicker from './TimePicker'
 
@@ -78,6 +78,14 @@ export const StyledTimePicker = styled(TimePicker)`
     display: inline-block;
     padding-top: 6px;
     vertical-align: middle;
+  }
+
+  &.small {
+    font-size: ${font.xl.size};
+
+    & > time {
+      padding-top: ${space.s};
+    }
   }
 `